add monthly community health report system
