+++ pontoon/base/static/css/style.css
@@ -787,11 +787,6 @@ header .right .select .menu {
   color: var(--white-1);
 }
 
-.notifications .menu li.notification-item .description ul {
-  list-style: disc;
-  padding: 10px 0 10px 20px;
-}
-
 .notifications .menu li.notification-item .message {
   padding: 10px;
 }
@@ -819,13 +814,27 @@ header .right .select .menu {
   padding: 0;
 }
 
-/* Users can include HTML tags in their messages */
+.notifications .menu li.notification-item .description .positive {
+  color: var(--status-translated);
+}
+
+.notifications .menu li.notification-item .description .negative {
+  color: var(--status-error);
+}
+
+/* Notifications can include HTML tags in their body */
 .notifications .menu li.notification-item .message h1,
 .notifications .menu li.notification-item .message h2,
 .notifications .menu li.notification-item .message h3,
 .notifications .menu li.notification-item .message h4,
 .notifications .menu li.notification-item .message h5,
-.notifications .menu li.notification-item .message h6 {
+.notifications .menu li.notification-item .message h6,
+.notifications .menu li.notification-item .description h1,
+.notifications .menu li.notification-item .description h2,
+.notifications .menu li.notification-item .description h3,
+.notifications .menu li.notification-item .description h4,
+.notifications .menu li.notification-item .description h5,
+.notifications .menu li.notification-item .description h6 {
   color: var(--white-1);
   font-size: 14px;
   font-style: normal;
@@ -834,23 +843,28 @@ header .right .select .menu {
   text-transform: none;
 }
 
-.notifications .menu li.notification-item .message h1 {
+.notifications .menu li.notification-item .message h1,
+.notifications .menu li.notification-item .description h1 {
   font-size: 18px;
 }
 
-.notifications .menu li.notification-item .message h2 {
+.notifications .menu li.notification-item .message h2,
+.notifications .menu li.notification-item .description h2 {
   font-size: 16px;
 }
 
-.notifications .menu li.notification-item .message p {
+.notifications .menu li.notification-item .message p,
+.notifications .menu li.notification-item .description p {
   padding: 10px 0;
 }
 
-.notifications .menu li.notification-item .message ol {
+.notifications .menu li.notification-item .message ol,
+.notifications .menu li.notification-item .description ol {
   padding-left: 20px;
 }
 
-.notifications .menu li.notification-item .message ul {
+.notifications .menu li.notification-item .message ul,
+.notifications .menu li.notification-item .description ul {
   list-style: disc;
   padding: 10px 0 10px 20px;
 }
