+++ pontoon/contributors/templates/contributors/widgets/notifications_menu.html
@@ -87,7 +87,7 @@
           {% endif %}
 
           {% if description and (description.startswith("Unreviewed suggestions") or notification.verb == "has reviewed suggestions" or notification.verb == "ignore") %}
-            <span class="description">{{ description|safe }}</span>
+            <div class="description">{{ description|safe }}</div>
 
             {{ notification_time(notification) }}
           {% elif target and target.resource %}
